+++ CHANGELOG.rst
@@ -10,11 +10,22 @@ The format is based on `Keep a Changelog`_ and this project adheres to `Semantic
 
 1.0.0 (UNRELEASED)
 ==================
+
+Added
+-----
 - Add a lot of documentation
+
+Changed
+-------
+- Rework exceptions to make things more straightforward
 
 
 0.1.0 (2024-09-25)
 ==================
+
+Initial release
+---------------
+
 - Full implementation for ``LD2410C`` devices on firmware ``v2.04.23022511``
 - Linted and typed code, using ruff_ and mypy_
 - Tests with a dedicated ``LD2410`` emulator, code is fully covered